Test implementati e pulizia codice

+++ src/main/java/ingsoftware/model/enum_helpers/HabitDifficulty.java
@@ -1,19 +1,19 @@
 package ingsoftware.model.enum_helpers;
 
 public enum HabitDifficulty {
-    EASY(10, 1.2),
-    MEDIUM(15, 1.0),
-    HARD(20, 0.8);
+    EASY( 10.0, 1.2),
+    MEDIUM(15.0, 1.0),
+    HARD(20.0, 0.8);
 
-    private final int baseXP;
+    private final double baseXP;
     private final double penaltyMultiplier;
 
-    HabitDifficulty(int baseXP, double penaltyMultiplier) {
+    HabitDifficulty(double baseXP, double penaltyMultiplier) {
         this.baseXP = baseXP;
         this.penaltyMultiplier = penaltyMultiplier;
     }
 
-    public int getBaseXP() {
+    public double getBaseXP() {
         return baseXP;
     }
